refactor(#3479): extract provider-session persistence from tmux_watcher.rs
Verbatim move (zero logic change) of resolve_persistable_provider_session_id + persist_watcher_provider_session_id (and their 3 unit tests) from tmux_watcher.rs into the new sibling tmux_watcher/provider_session_persistence.rs (101 prod LoC); tmux_watcher.rs drops 7022 -> 6939 prod LoC, ratchet baseline lowered to match.
